+++ loris/webapp.py
@@ -9,6 +9,8 @@ from __future__ import absolute_import
 
 from datetime import datetime
 from decimal import getcontext
+import logging
+from logging.handlers import RotatingFileHandler
 import os
 from os import path, unlink
 import re
@@ -23,7 +25,6 @@ from werkzeug.wrappers import (
 )
 
 from loris import constants, img, transforms
-from loris.config import configure_logging
 from loris.img_info import ImageInfo, InfoCache
 from loris.loris_exception import (
 	ConfigError,
@@ -90,6 +91,92 @@ def read_config(config_file_path):
     config['DEFAULT'] = {key: val for(key, val) in os.environ.items() if key not in ('PS1')}
     return config
 
+
+def _validate_logging_config(config):
+    """
+    Validate the logging config before setting up a logger.
+    """
+    mandatory_keys = ['log_to', 'log_level', 'format']
+    missing_keys = []
+    for key in mandatory_keys:
+        if key not in config:
+            missing_keys.append(key)
+
+    if missing_keys:
+        raise ConfigError(
+            'Missing mandatory logging parameters: %r' %
+            ','.join(missing_keys)
+        )
+
+    if config['log_to'] not in ('file', 'console'):
+        raise ConfigError(
+            'logging.log_to=%r, expected one of file/console' % config['log_to']
+        )
+
+    if config['log_to'] == 'file':
+        mandatory_keys = ['log_dir', 'max_size', 'max_backups']
+        missing_keys = []
+        for key in mandatory_keys:
+            if key not in config:
+                missing_keys.append(key)
+
+        if missing_keys:
+            raise ConfigError(
+                'When log_to=file, the following parameters are required: %r' %
+                ','.join(missing_keys)
+            )
+
+
+def configure_logging(config):
+    _validate_logging_config(config)
+
+    logger = logging.getLogger()
+
+    try:
+        logger.setLevel(config['log_level'])
+    except ValueError:
+        logger.setLevel(logging.DEBUG)
+
+    formatter = logging.Formatter(fmt=config['format'])
+
+    if not getattr(logger, 'handler_set', None):
+        if config['log_to'] == 'file':
+            fp = '%s.log' % (path.join(config['log_dir'], 'loris'),)
+            handler = RotatingFileHandler(fp,
+                maxBytes=config['max_size'],
+                backupCount=config['max_backups'],
+                delay=True)
+            handler.setFormatter(formatter)
+            logger.addHandler(handler)
+        else:
+            from sys import __stderr__, __stdout__
+            # STDERR
+            err_handler = logging.StreamHandler(__stderr__)
+            err_handler.addFilter(StdErrFilter())
+            err_handler.setFormatter(formatter)
+            logger.addHandler(err_handler)
+
+            # STDOUT
+            out_handler = logging.StreamHandler(__stdout__)
+            out_handler.addFilter(StdOutFilter())
+            out_handler.setFormatter(formatter)
+            logger.addHandler(out_handler)
+
+            logger.handler_set = True
+    return logger
+
+
+class StdErrFilter(logging.Filter):
+    '''Logging filter for stderr
+    '''
+    def filter(self,record):
+        return 1 if record.levelno >= 30 else 0
+
+class StdOutFilter(logging.Filter):
+    '''Logging filter for stdout
+    '''
+    def filter(self,record):
+        return 1 if record.levelno <= 20 else 0
 
 class LorisResponse(BaseResponse, CommonResponseDescriptorsMixin):
     '''Similar to Response, but IIIF Compliance Link and
